Refactor session actions into dedicated controllers
