+++ ext_localconf.php
@@ -2,6 +2,10 @@
 declare(strict_types=1);
 
 use Ndrstmr\Dt3Pace\Controller\SessionController;
+use Ndrstmr\Dt3Pace\Controller\SessionListController;
+use Ndrstmr\Dt3Pace\Controller\SessionProposalController;
+use Ndrstmr\Dt3Pace\Controller\SessionVoteController;
+use Ndrstmr\Dt3Pace\Controller\SessionApiController;
 use Ndrstmr\Dt3Pace\Controller\SpeakerController;
 use Ndrstmr\Dt3Pace\Controller\NoteController;
 use TYPO3\CMS\Extbase\Utility\ExtensionUtility;
@@ -9,12 +13,14 @@ use TYPO3\CMS\Extbase\Utility\ExtensionUtility;
 defined('TYPO3') or die();
 
 ExtensionUtility::configurePlugin('Ndrstmr.Dt3Pace', 'Agendalist', [
-    SessionController::class => 'list,show'
+    SessionListController::class => 'list',
+    SessionController::class => 'show'
 ], []);
 
 ExtensionUtility::configurePlugin('Ndrstmr.Dt3Pace', 'Agendagrid', [
-    SessionController::class => 'grid,listJson'
-], [SessionController::class => 'listJson']);
+    SessionListController::class => 'grid',
+    SessionApiController::class => 'listJson'
+], [SessionApiController::class => 'listJson']);
 
 ExtensionUtility::configurePlugin('Ndrstmr.Dt3Pace', 'Sessionshow', [
     SessionController::class => 'show'
@@ -29,12 +35,13 @@ ExtensionUtility::configurePlugin('Ndrstmr.Dt3Pace', 'Speakershow', [
 ], []);
 
 ExtensionUtility::configurePlugin('Ndrstmr.Dt3Pace', 'Sessionform', [
-    SessionController::class => 'new,create'
-], [SessionController::class => 'create']);
+    SessionProposalController::class => 'new,create'
+], [SessionProposalController::class => 'create']);
 
 ExtensionUtility::configurePlugin('Ndrstmr.Dt3Pace', 'Sessionvoting', [
-    SessionController::class => 'listProposals,vote'
-], [SessionController::class => 'vote']);
+    SessionProposalController::class => 'listProposals',
+    SessionVoteController::class => 'vote'
+], [SessionVoteController::class => 'vote']);
 
 ExtensionUtility::configurePlugin('Ndrstmr.Dt3Pace', 'Eventsummary', [
     NoteController::class => 'summary'
@@ -42,14 +49,14 @@ ExtensionUtility::configurePlugin('Ndrstmr.Dt3Pace', 'Eventsummary', [
 
 $GLOBALS['TYPO3_CONF_VARS']['FE']['ajaxRoutes']['dt3pace_session_vote'] = [
     'path' => '/dt3pace/session/vote',
-    'target' => SessionController::class . '::voteAction',
+    'target' => SessionVoteController::class . '::voteAction',
     'access' => 'public',
     'methods' => ['POST'],
 ];
 
 $GLOBALS['TYPO3_CONF_VARS']['FE']['ajaxRoutes']['dt3pace_sessions_json'] = [
     'path' => '/dt3pace/sessions/json',
-    'target' => SessionController::class . '::listJsonAction',
+    'target' => SessionApiController::class . '::listJsonAction',
     'access' => 'public',
 ];
